requirements.txt
dill
Flask
gunicorn
Jinja2
joblib
json5
jsonschema
matplotlib
numpy
pandas
requests
scikit-learn
urllib3
xgboost-py